Clean up statuses
